+++ MelonLoader/Preferences/MelonPreferences_Category.cs
@@ -168,7 +168,7 @@ namespace MelonLoader
             }
             if (printmsg)
                 MelonLogger.Msg($"MelonPreferences Saved to {currentfile.FilePath}");
-            MelonHandler.OnPreferencesSaved();
+            MelonHandler.OnPreferencesSaved(currentfile.FilePath);
         }
 
         public void LoadFromFile(bool printmsg = true)