+++ MelonLoader/Preferences/MelonPreferences_ReflectiveCategory.cs
@@ -110,7 +110,7 @@ namespace MelonLoader.Preferences
             }
             if (printmsg)
                 MelonLogger.Msg($"MelonPreferences Saved to {currentfile.FilePath}");
-            MelonHandler.OnPreferencesSaved();
+            MelonHandler.OnPreferencesSaved(currentfile.FilePath);
         }
 
         public void LoadFromFile(bool printmsg = true)